requirements.txt
arrow==1.3.0
Bio==1.7.0
biopython==1.78
ConfigArgParse==1.7
dgl_cu113==0.9.1
gensim==4.1.2
Logbook==1.7.0.post0
matplotlib==3.7.2
numpy==1.24.4
pandas==2.0.2
scikit_learn==1.0.1
seaborn==0.13.2
torch==1.12.0+cu113
tqdm==4.62.3
